#5: Multiline Swap
The spoiler-censoring function is failing because paragraphs in reddit don't work with a single spoiler tag. This is fixable by changing the double-newline paragraph separator with a two-space newline separator.

Adding a simple `replace()` method call to the outer call to put the comment response in resolves this issue in my tests.